Update DI extensions & builder to include `IDefinitionLoader`
